requirements.txt
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.14
langchain-experimental==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
openai==2.21.0
faiss-cpu==1.13.2
fastapi==0.129.0
uvicorn==0.40.0
sse-starlette==3.2.0
python-dotenv==1.2.1
pydantic==2.12.5
tiktoken==0.12.0
